# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  email: string;
  /** @nullable */
  name: string | null;
}

export interface AuthSession {
  accessToken: string;
  user: AuthUser;
}

export interface SignupInput {
  email: string;
  /** @minLength 6 */
  password: string;
  /** @minLength 1 */
  name: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface Link {
  id: number;
  shortId: string;
  shortUrl: string;
  originalUrl: string;
  /** @nullable */
  title: string | null;
  /** @nullable */
  description: string | null;
  clicks: number;
  createdAt: string;
  updatedAt: string;
}

export interface LinkInput {
  originalUrl: string;
  title?: string;
  description?: string;
  /**
     * @minLength 3
     * @maxLength 32
     * @pattern ^[a-zA-Z0-9_-]+$
     */
  shortId?: string;
}

export interface LinkUpdate {
  originalUrl?: string;
  title?: string;
  description?: string;
  /**
     * @minLength 3
     * @maxLength 32
     * @pattern ^[a-zA-Z0-9_-]+$
     */
  shortId?: string;
}

export interface LinkSummary {
  totalLinks: number;
  totalClicks: number;
  mostClicked: Link | null;
}

export interface ResolvedLink {
  originalUrl: string;
  /** @nullable */
  title?: string | null;
}