# lib/api-zod/src/generated/types/link.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Link {
  id: number;
  shortId: string;
  shortUrl: string;
  originalUrl: string;
  /** @nullable */
  title: string | null;
  /** @nullable */
  description: string | null;
  clicks: number;
  createdAt: Date;
  updatedAt: Date;
}