# lib/api-zod/src/generated/types/linkSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Link } from './link';

export interface LinkSummary {
  totalLinks: number;
  totalClicks: number;
  mostClicked: Link | null;
}