# lib/api-zod/src/generated/types/linkUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LinkUpdate {
  originalUrl?: string;
  title?: string;
  description?: string;
  /**
     * @minLength 3
     * @maxLength 32
     * @pattern ^[a-zA-Z0-9_-]+$
     */
  shortId?: string;
}