# lib/api-zod/src/generated/types/signupInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SignupInput {
  email: string;
  /** @minLength 6 */
  password: string;
  /** @minLength 1 */
  name: string;
}